% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_gdal.R
\name{call_gdal}
\alias{call_gdal}
\title{Call GDAL}
\usage{
call_gdal(input, method, cellsize)
}
\arguments{
\item{input}{input grid (raster)}

\item{method}{resampling method}

\item{cellsize}{output cell diminsion}
}
\value{
RasterLayer
}
\description{
Call GDAL
}
\keyword{internal}
