% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_output.R
\name{prep_output}
\alias{prep_output}
\title{Prepare Output Grid}
\usage{
prep_output(input, cellsize = 1000)
}
\arguments{
\item{input}{the input grid to resample}

\item{cellsize}{the target cellsize of the output grid (in same units as input grid)}
}
\value{
an `sf` object of polygons (grid)
}
\description{
Prepare Output Grid
}
