% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_from_vector.R
\name{raster.from.vector}
\alias{raster.from.vector}
\title{Raster from Vector}
\usage{
raster.from.vector(output)
}
\arguments{
\item{output}{An `sf` grid to be resampled to}
}
\value{
a single layer `raster`
}
\description{
This function maps a vector to an `sf` grid and then to a raster
}
