% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass_utility.R
\name{reclass}
\alias{reclass}
\title{Reclassify a distribution table by a count value}
\usage{
reclass(areal_per, output_count, no_data = 0)
}
\arguments{
\item{areal_per}{the areal percent of each cell covered by a class}

\item{output_count}{the number of target cells for each class in the output cell}

\item{no_data}{the no data value in the native dataset}
}
\value{
a vector
}
\description{
Reclassify a distribution table by a count value
}
