% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resampleData}
\alias{resampleData}
\title{Resample}
\usage{
resampleData(
  input,
  cellsize = 1000,
  method = "area",
  no_data = NA,
  seed = 10291991
)
}
\arguments{
\item{input}{a raster of fine resolution}

\item{cellsize}{cell resolution of output target grid. In units of input CRS}

\item{method}{the method for resampling, nearest neighbor (nn), majority rule (maj), Area-based (area), or raw area (rawarea)}

\item{no_data}{a value to be treated as NO_DATA. Default to NA}

\item{seed}{seed number for random class assignment in rawarea}
}
\value{
a resampled raster
}
\description{
resample an input raster to an output grid using three different methods
}
