% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.R
\name{rna}
\alias{rna}
\title{Random Allocation}
\usage{
rna(input, output)
}
\arguments{
\item{input}{input grid}

\item{output}{output grid}
}
\value{
a vector of classifications
}
\description{
Random Allocation
}
