% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_output_cells.R
\name{summarize.cells2}
\alias{summarize.cells2}
\title{Summarize Output Grid Cells}
\usage{
\method{summarize}{cells2}(input, output, no_data = NA)
}
\arguments{
\item{input}{an input grid of fine resolution data `raster`}

\item{output}{the output grid to resample to. `sf` object}

\item{no_data}{the value representing "NO DATA", default is NA}
}
\value{
a tibble
}
\description{
Summarizies the percentage of each output grid cell that is covered by each disiticnt class in the input grid
}
