% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass_utility.R
\name{reclass}
\alias{reclass}
\title{Reclassify a Distribution Table by an Count value}
\usage{
reclass(areal_per, output_count, no_data = 0)
}
\arguments{
\item{areal_per}{the aeral percent of each cell covered by a class}

\item{output_count}{the number of target cells for each class in the output cell}

\item{no_data}{a value to set unclassified cells to}
}
\value{
a vector
}
\description{
Reclassify a Distribution Table by an Count value
}
