% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resampleData}
\alias{resampleData}
\title{Resample}
\usage{
resampleData(input, output, no_data = NA, method = "area")
}
\arguments{
\item{input}{a raster of fine resolution}

\item{output}{an output grid (can be created with \code{prep_output})}

\item{no_data}{a value to be treated as NO_DATA. Default to NA}

\item{method}{the method for resampling, nearest neighbor (nn), majority rule (maj) or Area-based (area)}
}
\value{
a resampled raster
}
\description{
resample an input raster to an output grid using three different methods
}
